Improve FileDeclaration in the Reactor
